+++ src/components/molecules/vTaskItemTeacher.vue
@@ -84,7 +84,14 @@ export default {
 			return this.task.status.isDraft;
 		},
 		courseName() {
-			return this.task.courseName || this.$t("pages.tasks.labels.noCourse");
+			const baseName =
+				this.task.courseName || this.$t("pages.tasks.labels.noCourse");
+			const prefix =
+				this.task.status.isSubstitutionTeacher === true
+					? this.$t("common.words.substitute") + " "
+					: "";
+
+			return `${prefix}${baseName}`;
 		},
 		topic() {
 			return this.task.description