+++ src/components/templates/TasksDashboardTeacher.unit.js
@@ -1,21 +1,20 @@
 import TasksDashboardTeacher from "./TasksDashboardTeacher";
 import TasksList from "@components/organisms/TasksList";
 import vCustomEmptyState from "@components/molecules/vCustomEmptyState";
-import {
-	overDueTasksTeacher,
-	dueDateTasksTeacher,
-	noDueDateTasksTeacher,
-} from "@@/stories/mockData/Tasks";
+import mocks from "@@/stories/mockData/Tasks";
 import Vuetify from "vuetify";
+
+const { overDueTasksTeacher, dueDateTasksTeacher, noDueDateTasksTeacher } =
+	mocks;
 
 describe("@components/templates/TasksDashboardTeacher", () => {
 	const mockStore = {
 		tasks: {
 			getters: {
 				getStatus: () => "completed",
-				hasNoTasks: () => false,
-				hasNoOpenTasksTeacher: () => false,
-				hasNoDrafts: () => true,
+				hasTasks: () => true,
+				hasOpenTasksTeacher: () => true,
+				hasDrafts: () => false,
 				getOpenTasksForTeacher: () => ({
 					overdue: overDueTasksTeacher,
 					withDueDate: dueDateTasksTeacher,